Fix bug in GOSUB error handling
Prior to this fix cmd_gosub() would setup the interpreter state for
the GOSUB/RETURN before determining whether the target label/line
existed. If the target did not exist (an error occurred) then the
erroneous state would remain.

After this fix the state is not changed until the target label/line
has been confirmed to exist.