Rework CMM2/Game*Mite/MMB4W/PicoMiteVGA surface creation
OPTION SIMULATE device$ now always sets MODE 1 and creates the
surfaces for the new device, but the Display/Window is created
hidden and will not be shown until it is first rendered to.

Likewise setting the MODE will destroy existing surfaces 0..63
and recreate them with a new hidden Display/Window.