Support multibyte function tokens
 ... thus removing the limit of 127/128 function tokens.

Note that FunctionToken values are not indexes into the tokentbl[],
you have to add 128 to them. This is different to CommandToken where
the values are indexes into commandtbl[]. This difference in the
behaviour of FunctionToken allows the non-token characters 0-127 to
be read and written with fewer special cases.

The default behaviour is for FunctionTokens up to and including 254
(0xFE) to be represented by a single-byte token whereas 255+ are
represented by THREE bytes, the first of which is 0xFF, see
tokentbl.h for details.

By uncommenting the definition of USE_TWO_BYTE_TOKENS in tokentbl.h
you can switch the code to using TWO bytes for all FunctionTokens
with values > 127.

It is my intention to make the encoding of CommandTokens consistent
with FunctionTokens in the near future.