feat: change platform switching behaviour
If a platform is supplied at the command line then it is used as the
default (instead of MMB4L) that is returned to by the NEW and RUN
commands.

In addition the NEW and RUN commands no longer destroy the graphics
surfaces being used by the display/console unless the platform is
switched.