refactor: move/rename MMgetchar() => prompt_getc()
Also engaged in some AI assisted "change & pray" on the handling
of input from a pipe/file being exhausted.

This is untested because currently MMBasic cannot take input from
a pipe/file due to how it trys to determine the TTY cursor
position by writing ANSI escape to STDOUT and then expects a
response on STDIN - which can't happen when STDIN is a pipe/file.

Possibly if a kStdinExhausted is encountered by the error
handling code the program should exit after printing the error
instead of performing a longjmp().